chore(dispatch): replace catch-all unreachable with exhaustive match; unify clap error formatting (#1368) (#1374)
## Test plan

- **Exhaustiveness test**: a compile-time check that every `Commands`
variant appears in either `dispatch.rs` or `rest.rs`, with no catch-all.
The `#[deny(unreachable_patterns)]` lint gives this for free.
- **Integration**: `podup logs --wrong-flag` exits non-zero with the
`podup:` prefix.
- **Unit test** for `is_label_only` future-proofing (covered here as
part of the same shape concern): a `#[non_exhaustive]` annotation on the
enum would catch it at compile time.

---------
